+++ src/dotnet/Blazor3D/Blazor3D/Geometires/PlaneGeometry.cs
@@ -18,7 +18,7 @@ namespace Blazor3D.Geometires
         /// <param name="height">The length of the edges parallel to the Y axis. Default is 1.</param>
         /// <param name="widthSegments">Number of segmented rectangular faces along the width of the sides. Default is 1.</param>
         /// <param name="heightSegments">Number of segmented rectangular faces along the height of the sides. Default is 1.</param>
-        public PlaneGeometry(float width = 1, float height = 1, int widthSegments = 1, int heightSegments = 1) : this()
+        public PlaneGeometry(double width = 1, double height = 1, int widthSegments = 1, int heightSegments = 1) : this()
         {
             Width = width;
             Height = height;
@@ -29,12 +29,12 @@ namespace Blazor3D.Geometires
         /// <summary>
         /// The length of the edges parallel to the X axis. Default is 1.
         /// </summary>
-        public float Width { get; set; } = 1;
+        public double Width { get; set; } = 1;
 
         /// <summary>
         /// The length of the edges parallel to the Y axis. Default is 1.
         /// </summary>
-        public float Height { get; set; } = 1;
+        public double Height { get; set; } = 1;
 
         /// <summary>
         /// Number of segmented rectangular faces along the width of the sides. Default is 1.
