+++ src/dotnet/Blazor3D/Blazor3D/Geometires/TorusKnotGeometry.cs
@@ -21,7 +21,7 @@ namespace Blazor3D.Geometires
         /// <param name="tubularSegments">Number of tubular segments. Default is 64.</param>
         /// <param name="p">This value determines, how many times the geometry winds around its axis of rotational symmetry. Default is 2.</param>
         /// <param name="q">This value determines, how many times the geometry winds around a circle in the interior of the torus. Default is 3.</param>
-        public TorusKnotGeometry(float radius = 1, float tube = 0.4f, int radialSegments = 8, int tubularSegments = 64, int p = 2, int q = 3) :this()
+        public TorusKnotGeometry(double radius = 1, double tube = 0.4f, int radialSegments = 8, int tubularSegments = 64, int p = 2, int q = 3) :this()
         {
             Radius = radius;
             Tube = tube;
@@ -35,12 +35,12 @@ namespace Blazor3D.Geometires
         /// <summary>
         /// Radius of the torus, from the center of the torus to the center of the tube. Default is 1.
         /// </summary>
-        public float Radius { get; set; } = 1;
+        public double Radius { get; set; } = 1;
 
         /// <summary>
         /// Radius of the tube. Default is 0.4.
         /// </summary>
-        public float Tube { get; set; } = 0.4f;
+        public double Tube { get; set; } = 0.4f;
 
         /// <summary>
         /// Number of radial segments. Default is 8.
