+++ src/dotnet/Blazor3D/Blazor3D/Materials/MeshStandardMaterial.cs
@@ -32,12 +32,12 @@ namespace Blazor3D.Materials
         /// <summary>
         /// <para>How much the material is like a metal. Non-metallic materials such as wood or stone use 0.0, metallic use 1.0, with nothing (usually) in between. Default is 0.0. A value between 0.0 and 1.0 could be used for a rusty metal look. If metalnessMap is also provided, both values are multiplied.</para>
         /// </summary>
-        public float Metalness { get; set; } = 0;
+        public double Metalness { get; set; } = 0;
 
         /// <summary>
         /// <para>How rough the material appears. 0.0 means a smooth mirror reflection, 1.0 means fully diffuse. Default is 1.0. If roughnessMap is also provided, both values are multiplied.</para>
         /// </summary>
-        public float Roughness { get; set; } = 1;
+        public double Roughness { get; set; } = 1;
 
         /// <summary>
         /// <para>Render geometry as wireframe. Default is false (i.e. render as flat polygons).</para>
